requirements.txt
pillow==11.2.1
python-sane==2.9.1
